requirements.txt
asgiref==3.7.2
Django==4.2.7
django-phonenumber-field==7.2.0
install==1.3.5
phonenumbers==8.13.26
psycopg2-binary==2.9.9
sqlparse==0.4.4
tzdata==2023.3
